Commu Part/requirements.txt
dash
dash.dependencies
pandas
plotly.graph_objs
